+++ src/ncc/Classes/NccExtension/PackageCompiler.php
@@ -31,7 +31,6 @@
     use ncc\Classes\ComposerExtension\ComposerSourceBuiltin;
     use ncc\Classes\PhpExtension\PhpCompiler;
     use ncc\CLI\Main;
-    use ncc\Exceptions\AccessDeniedException;
     use ncc\Exceptions\BuildException;
     use ncc\Exceptions\ConfigurationException;
     use ncc\Exceptions\IOException;
@@ -58,7 +57,6 @@
          * @param ProjectManager $manager
          * @param string $build_configuration
          * @return string
-         * @throws AccessDeniedException
          * @throws BuildException
          * @throws ConfigurationException
          * @throws IOException
@@ -155,7 +153,6 @@
          * @param string $path
          * @param ProjectConfiguration $configuration
          * @return array
-         * @throws AccessDeniedException
          * @throws IOException
          * @throws PathNotFoundException
          * @throws RunnerExecutionException