+++ src/ncc/Classes/PythonExtension/PythonRunner.php
@@ -22,7 +22,6 @@
 
     namespace ncc\Classes\PythonExtension;
 
-    use ncc\Exceptions\AccessDeniedException;
     use ncc\Exceptions\IOException;
     use ncc\Exceptions\PathNotFoundException;
     use ncc\Interfaces\RunnerInterface;
@@ -38,7 +37,6 @@
          * @param string $path
          * @param ExecutionPolicy $policy
          * @return ExecutionUnit
-         * @throws AccessDeniedException
          * @throws IOException
          * @throws PathNotFoundException
          */